+++ package.json
@@ -1,27 +1,19 @@
 {
-  "name": "react-responsive",
-  "version": "0.1.0",
-  "private": true,
+  "name": "@doux/react-responsive",
+  "version": "0.0.1",
+  "source": "src/react-responsive.tsx",
+  "main": "dist/react-responsive.jsx",
+  "module": "dist/react-responsive.mjsx",
+  "unpkg": "dist/react-responsive.jsx",
+  "types": "dist/react-responsive.tsx",
+  "amdName": "ReactResponsive",
   "scripts": {
-    "build": "react-scripts build"
+    "build": "microbundle --no-compress --jsx --globals react=React --strict",
+    "test":"echo \"Error: no test specified\" && exit 1"
   },
-  "eslintConfig": {
-    "extends": [
-      "react-app",
-      "react-app/jest"
-    ]
-  },
-  "browserslist": {
-    "production": [
-      ">0.2%",
-      "not dead",
-      "not op_mini all"
-    ],
-    "development": [
-      "last 1 chrome version",
-      "last 1 firefox version",
-      "last 1 safari version"
-    ]
+  "repository": {
+    "type": "git",
+    "url": "git+https://github.com/ZzhangFun/react-response.git"
   },
   "devDependencies": {
     "@types/jest": "^28.1.3",
@@ -33,5 +25,9 @@
     "react-dom": "^18.2.0",
     "typescript": "^4.7.4",
     "web-vitals": "^2.1.4"
-  }
+  },
+  "peerDependencies": {
+    "react": ">=16.8"
+  },
+  "files": ["dist", "README.md"]
 }
